feat(chat): 소켓 연결 시작
