+++ src/pages/chat/apis/chat-ws.api.ts
@@ -1,10 +1,9 @@
-import { useState, useEffect } from 'react';
-
 import { subscribe, sendMessage } from '@shared/service';
 
 // 사용자 입장 알림
 export const joinChatRoom = (userName: string, roomName: string) => {
   sendMessage('/api/app/chat/join', { userName, roomName });
+  console.log(`[${userName}] 님이 [${roomName}] 채팅방에 입장했습니다.`);
 };
 
 // 메시지 전송
@@ -16,27 +15,9 @@ export const sendChatMessage = (roomName: string, sender: string, content: strin
 export const subscribeUsers = (
   callback: (users: { userName: string; roomName: string }[]) => void,
 ) => {
-  subscribe('/topic/users', callback);
+  return subscribe('/topic/users', callback); // 구독 객체 반환
 };
 
-// 채팅방별 접속인원 목록 구독
-export const SubscribeRoomUsers = () => {
-  const [roomUsers, setRoomUsers] = useState<{ [key: string]: number }>({});
-
-  useEffect(() => {
-    const subscription = subscribe('/topic/room-users', (data) => {
-      setRoomUsers(data);
-    });
-
-    return () => {
-      if (subscription) {
-        subscription.unsubscribe();
-      }
-    };
-  }, []);
-
-  return roomUsers;
-};
 // 특정 채팅방 메시지 구독
 export const subscribeRoomMessages = (
   roomName: string,
