+++ src/pages/chat/hooks/index.ts
@@ -2,3 +2,4 @@ export * from './useGetChatRooms';
 export * from './useChatRoomContext';
 export * from './useCreateChatRoom';
 export * from './useGetChatMessages';
+export * from './useSubscribeRoomUsers';
