+++ src/pages/chat/ui/ChatPage.tsx
@@ -6,37 +6,38 @@ import { Box, Flex, Button } from '@chakra-ui/react';
 import { RouterPath } from '@shared/constants';
 import { connectWebSocket, disconnectWebSocket } from '@shared/service';
 
-import { joinChatRoom, subscribeUsers, subscribeRoomMessages, SubscribeRoomUsers } from '../apis';
+import { joinChatRoom } from '../apis';
 import { ChatRoomSection, ChatRoomInsideSection, ChatUserSection } from '../components';
 import { useChatRoomContext } from '../hooks';
 
 export const ChatPage = () => {
   const navigate = useNavigate();
   const userNickname = localStorage.getItem('userNickname') || 'Guest';
-  const { selectedRoom } = useChatRoomContext();
+  const { selectedRoom, setSelectedRoom } = useChatRoomContext();
+
+  useEffect(() => {
+    if (!selectedRoom) {
+      setSelectedRoom('미접속');
+    }
+    // eslint-disable-next-line react-hooks/exhaustive-deps
+  }, []);
 
   // 웹소켓 연결 및 구독
   useEffect(() => {
-    if (selectedRoom) {
-      connectWebSocket(
-        () => {
-          console.log('WebSocket 연결됨');
-          joinChatRoom(userNickname, selectedRoom);
-          SubscribeRoomUsers();
-          subscribeRoomMessages(selectedRoom, (message) => {
-            console.log('새 메시지:', message);
-          });
-        },
-        (error) => {
-          console.error('WebSocket 연결 오류:', error);
-        },
-      );
-    }
+    connectWebSocket(
+      () => {
+        console.log('WebSocket 연결됨');
+        joinChatRoom(userNickname, '미접속');
+      },
+      (error) => {
+        console.error('WebSocket 연결 오류:', error);
+      },
+    );
 
     return () => {
       disconnectWebSocket();
     };
-  }, [selectedRoom, userNickname]);
+  }, [userNickname]);
 
   return (
     <Flex flexDir='column' alignItems='center' w='full' h='full'>
